cellstr fill or clip text as required
If the amount of text to display for a cell is greater than the cell
width, then clip the text to match the width.

If the amount of text to display for a cell is smaller than the cell
width, then fill it with spaces to match the width.

The actual text data is not affected, only the display of text is
affected.