Plot nr and nc cells, move cursor left right
Previously I was plotting one row and one col less than the specified
number of rows and cols (rather it was assumed to be the 0th row and
col), now instead I actually plot the specified number of rows and cols
independent of the 0th row and 0th col, which correspond to matrix
marker/header areas).

Allow the current cursor to be moved along horizontal direction, i.e
left and right.