cellcur_left/right adjust viewport, cdraw just use it
Now cellcur left and right logic adjust the viewport as required.
cdraw inturn just uses the same.

Have trapped exception from runlogic and print the same to stderr.

Similarly if any debugging required the same could be written to
stderr. Inturn when starting the program redirect the stderr to a
file and look at the file to see what is going on.