Changed default mode wrt displaying cell matrix
Now the printing of cell address to help with debugging has been
made optional.

Also now by default the cell content are not clipped. So if the
adjacent cell doesnt have any data to display, the cell content
is not clipped, else it will be clipped by adjacent cell data.