Fix cell overflow display, update data only after enter key
THe previous logic was not proper when cell content was overflowing.

This new logic takes care of showing cells with overflowing data as
required be it in clip mode or non clip mode.

In clip mode, any overflowing text is not displayed.

In non clip mode, the overflowing text is displayed, if there is no
text/data in adjacent cell. If there is text in adjacent cell, then
the overflowing text is not displayed.

The data being entered in insert mode is maintained in the gotStr.
And the cdrawdata now uses the data from gotStr as required, so there
is no longer any need to update the actual data dictionary with the
same temporary data.