Take care of editing buffer
If entering edit mode, then the existing content of the cell is
backed up and inturn as the edit progresses, if enter key is pressed
the new updated edit buffer is backed up.

Inturn when exiting into command mode. THe data in the backedup uptodate
edit buffer is saved into the data dictionary wrt that cell.

Also parallely the data in the data dictionary wrt the cell is cleared,
when entering edit mode, which is fine, because the backedup edit buffer
keeps track of the data as required.