CellValidityCheck returns cellKey for data dict
Now the celladdr_valid not only checks if the given string is a
valid cell address or not. It also returns the cellKey for the
data dictionary, if the given string is a valid cell address.

nvalud_saddr which is not used with this new flow and semantic, is
not implemented. It logs a error and exits the program, if called.