loadfile: handle any field data at the end of line
Previously field contents of a csv file would have been saved, only
if it is followed by a field seperator.

However to be on safe side one should expect that there could be
data as part of the last field in a given row, but then not terminated
by a field seperator, because it is the last field and so will naturally
be terminated by a new line.

This logic now takes care of such last field data.