Clear screen when coming back into Command mode
While entering text into a cell, one may go beyond the right side
end of the screen. In which case when coming back into command mode
there could be some stray appearing characters to the right most end
of the screen corresponding to the edit cell row, in some cases,
especially because the command mode doesnt fill the full screen in
many cases, based on the size of the screen. And also sometimes
based on the font used.

This new logic takes care of this situation.