Added status, updated dlg, use in save_file
Added a status func which will show a list of messages passed to it.

Both status and updated dlg implicitly show the list of messages
passed to them at the top right corner of the screen specified, ie
if no position is specified explicitly. Also they use normal attribute
if it is not explicitly specified.

	dlg arguments order changed to accomodate above wrt default
	args (i.e args with default values).

save_file uses status func to provide status, based on user actions.