Use a exit key to trigger a controlled exit from program
This ensures that any unwinding or restoring can be handled as part
of the normal flow, follwed by a graceful exit from the program.

Quit logic checks there is no unsaved changes before triggering the
exit. If unsaved changes exists, then it asks the user whether to
exit or not.