Implicit stderr redir failed, Explicit file setting
For some reason forcing stderr to a named temporary file by directly
assigning to sys.stderr failed. Need to look at the reason and do it
the right way later.

So for now I create a global file handle for errors, which is initialed
using a named temporary file and inturn make print_exc write into this file.

As temporary file is opened in raw/byte mode by default, so explicitly
set the mode to text.