Forgot that curses refresh is required to update display
As I havent used curses in ages and then did a quick scan through
some of the docs when starting on this few days back, so I forgot
about the need to refresh to get things onto the display/terminal.

And getch automatically refreshes, so dlg msgs where visible and
status msgs wherenot.

THis is a commit which will be reverted, captured here only for
reference for future, in case I or anyone else goes through these.