Expand encrypt step to help debug
The str to bytes and or bytes to str shouldnt bring in any variations,
other than the type conversion, but just to be sure, given the variance
in encrypted data.

NOTE: Also come to think of it, I think while glancing through the info,
somewhere I read about timestamp being maintained, so potentially that
would be what is causing the variance. WIll be decrypting next to confirm
the data is fine and the flow I am doing is fine.