UpdateCellAddr:P5: delete_rc tries2Adjust celladdrs
Any cell address which references cell(s) which is beyond the deleted
row or col will have itself adjusted.

NOTE:TODO:TOTHINK: However any cell address which references the
deleted row or col, is left as it is. Need to think should I leave
it as it is, or should I change such cell addresses to a special
value or ...