UpdateCellAddrs:P6: cellAddrValid fix
Make cellAddrValid return similar set of items in its failure path
as it does in its valid path.

Previously as it was mainly called from locations where the address
passed to it for its operation would be valid cell address and or
places where it could valid or otherwise I tested mainly with valid
addresses, so this asymmetry didnt create any problem and got missed
out. This now is fixed.