Dirty flag wanted
Even thou I had added the concept of dirty flag yesterday or so,
I had not used it fully as required in all needed paths.

Have fixed it partly by now setting dirty flag if one pastes and
or cuts some thing. Some minimal check is done to ensure that it
gets set only if there is some non-imaginary ;_) data ;-)

Have also added same for insert_rc_ab and delete_rc path. This is
a blind setting of the flag at some levels, in that if these cmds
are triggered and if the correponding helper functions return
without raising any exception, then the state is docked into dirty
path.