Set crsrOffset in editplus mode
Now the logic sets the crsrOffset in the following situations of
the editplus mode:

	* When entering edit mode
	* When adding characters to current edit buffer
	* when removing characters using backspace from edit buffer

Inturn cellcur updated to account for the crsrOffset, when in edit
mode (that also includes the insert command triggered edit mode).

NOTE: On entering edit mode, the cursor is enabled, only after the
runlogic gets the next key press, so cursor wont show till user
presses something in edit mode.

NOTE: Currently as one can only edit by adding to or removing from
the end of the current edit buffer string, so the logic for now
inefficiently and lazily recalculates the current edit buffer string
(i.e gotStr) and sets up the crsrOffset each time a char is added or
removed (i.e instead of inc or dec the current crsrOffset).