explicit command mode cursor
Now cellcur updated to handle showing of cursor at right location
for the ':' explicit command mode. Here we require to add +1 to
crsrOffset to get the actual cursor position. This is because,
there is a implicit ':' char at the begining in explicit command
mode.

cdraw updated to call cellcur wrt 0,0 cell when in ':' explicit
command mode.