Ignore arrow keys in editplus mode
THis ensures that if the user presses the arrow keys, then it is
ignored rather than special characters getting inserted into the
edit buffer.

This also ensures that the current cursor position at the end of
the edit buffer is shown. So the situation mentioned in a prev
commit wrt on entering edit mode not showing the blinking text
cursor till user presses some key is also taken care of in a indirect
sense in that, when user presses arrow keys, even thou the key
is ignored, the runlogic will switch states and enable cursor mode
and thus start showing the edit text cursor location.

NOTE: using isprintable doest help here, because the integer code
used by curses for arrow keys map to valid characters at a different
level, so cant help differentiate. SO the need to use a specific
list of codes like done in this commit.