Moving beyond screen - trap wmove
In editing mode, if one enters a very long line, which goes beyond
the screen wrt current line, then it wraps around to next line and
continuous from the left edge of this new line.

However this also parallely triggers a wmove error/exception from
curses. This exception is trapped now.

It is noticed that if one traps this exception, the end user is not
impacted in any other way and they can edit long lines which wrap
into next line during editing.

If we dont trap this exception, then the program will exit due to
exception.