Cursor position in overflowed lines
If the edit buffer overflows beyond its original line, then now the
cursor can be clearly positioned in any location in these overflowed
lines.

Previously it could be only positioned at the end of the overflowed
line and not any inbetween position, unless it was a position in the
original line itself. This has been fixed now.