Scroll if editbuffer overflows beyond screen vertically
As noted previously, if the edit buffer data overflows such that
it goes beyond the bottom of the screen (i.e vertically)

    a) Because we are editing a cell in the last visible line on
       the screen

    b) Because we are editing a really really long long line such
       that it overflows into too many lines

the curses addstr will raise a error.

Now such a condition or any other exception as of now wrt cdraw
will be trapped and inturn the screen is scrolled by few lines,
to try and provide more space for the overflowing edit buffer.