Understand PKCS7 modules usage
Noticed that my pkcs7 padding logic wasnt getting unpadded by the
cryptography libraries pkcs7 unpadding logic.

SO changed to using cryptography libraries pkcs7 padding logic and
then noticed two things

    The padding wasnt filling the blocksize, in the debug print
    Also the aes encryption func raised error about data not matching
    the keylength.

thats when I did the RTFM and realised that the size passed to PKCS7
library needs to be in bits and not bytes.