Fix the overambitious oversight wrt blocksize
In the overzealous ferver to make a generic logic which can adapt
block cipher based on symmetric algo keysize, forgot that keysize
need not necessarily map 1-to-1 and directly to the blocksize.

As AES 128, 192 and 256 keys basically map to 128bit blocksize,
with different number of internal rounds based on the keysize.

So have fixed the blocksize to 16bytes i.e 128bit now, as its
currently aes based.