cdraw trap and vertScroll adjust onlyFor EditMode
THe logic added to vertically scroll down by few rows so that one
can add more to the edit buffer, has been properly restricted to
now trigger this behaviour only in Edit mode.

Else if there is some error in cdraw due to cell expression handling
or printing, it would have forced scroll till the end of the spreadsheet
in the worst case.

The forcing of printing of cells beyond the screen by forcing cellwidth
to 2 tirggers this scenario.