set ERR/Logfile in CUI, full screen spreadsheet
Propogate the Error and Logfiles from main program to cui library.

Set cellwidth based check in cellstr to 0, so that it passes the
string to pass into curses, irrespective of message or cellwidth.
This ensures that even partial cell is shown on the screen, thus
using full space on the screen.

Inturn to achieve the same the cui libraries cellstr is called,
instead of directly calling addstr of curses. CUI's cellstr takes
care of clipping the message to screen if requested.

However the new clipToScreen argument should be passed to cellstr
in the spreadsheet now. Currently because this is not done, in the
edit mode and explicit command mode if the edit buffer overflows,
then the overflown data is not shown, because the clipToScreen is
enabled by default and which is required for normal spreadsheet
usage.