Ensure edit/explicitBuffer shownEvenIF dataCrashes
Previously if printing/showing data cells where to crash/raise a
exception, then even thou the exception is trapped, the display
of the edit / explicit command mode edit buffer would have got
bypassed, thus user not being able to see what they are typing
in to try and recover from such bad data or ...

THis has been fixed now.

At same time if the edit buffer overflows beyond the screen, the
logic will still raise a exception and allow the cdraw's exception
trapper to trap it and scroll down the display contents to adapt
to the overflowing edit buffer situation.