multi row/col delete_rc - P2
Adjust numRows/numCols as required.

If user doesnt give any argument to dr or dc explicit command,
force the arg to 1, so that the current row or col is deleted,
which will mimic the old behaviour of dr and dc as well as is a
safe default to achieve a simple delete row/col command.