Multi row/col delete_rc - P3
If trying to delete more than the available rows/cols, then ensure
that it identifies this situation and handle it properly.

Previously it would have eaten into rows/cols adjacent to current
row / col in the wrong direction, related to the amount of deficit
in the rows/cols to delete.