UpdateCellAddr Tags Err for row/col deleted
If a cell address reference refers to a deleted row or column
directly, then its prefixed with a ErrR or ErrC tag, so that
the user can look at the =expression containing this Err tagged
cell address and adjust suitabely, as the user finds right.

README changelog updated wrt this.