Updated test.csv to use round and pow
TODO: As cell values get evaluated as floats in some cases, so
round cant specify the numOfDecimalDigits argument as a cell value
currently.

At same time I cant call int function on the cell value to get a
int value and then pass it as the numOfDecimalDigits argument to
round or any other function requiring int value as its argument,
because the _nvalue_funcs currently parses for functions in a dumb
way.