Dont force eval result to float
If one wants to use round or any other python function, which may
require a int argument, which inturn is refered from a cell, then
previously it would have failed (resulted in None) because the
cell result was forced to be a float always.

Now instead the logic allows the eval return type to be used as is
rather than being force converted to float.