Avoid a exception corner case
If terminal size changes when one is waiting on getch, then
the returned value will not map to a valid character, but goes
beyond the valid range of chars and this raises a exception.

Avoid this by not tryint to convert to char, instead convert the
char to be checked into a interger value and then check.