load_help:P1: load help.csv in readonly mode
If the load_help has not already loaded, then load the help.csv
file. And inturn set the readOnly mode to True after saving its
current state into helpModeSavedReadOnly.

WHile loading if helpModeSavedReadOnly is not None, then on successful
loading set the readONly mode to value in helpModeSavedReadOnly.

TODO: Currently in load_help we cant be sure if loading of help.csv
was successful or not (because load_file traps any exception that
may be raised). So load_help can wrongly set the helpModeSavedReadOnly
state, even if the loading of help.csv was a failure. This inturn will
not allow help file to be loaded in future, till the program quits, bcas
load_help checks for helpModeSavedReadONly to be None before trying to
load the help.csv file. This needs to be fixed.