load_help:P2: Handle load_file failure appropraitely
NOTE: Had forgotten to name helpMode variable as helpModeSavedReadOnly,
even thou commit was refering to this name. This is fixed in this commit.

Also Now load_file returns true or false based on whether it was
successful or not. Inturn load_help checks this return value to decide,
whether it should save the current readOnly state and then inturn set
the readOnly mode to True or not.