Make GetTokenCellAddr parsekvc based - P1
Removed get_token which is used to get cell addresses, from
spreadsheetkvc.

Inturn have put the skeleton for a new get_celladdr in parsekvc,
which is based on helper routines in it.

There are few aspects of the logic which is yet to be done, as
captured in the TODO.

Also need to cross-check and adjust the logic in spreadsheetkvc,
which was using the get_token(...celladdrType) routine.