ParseKVC: get_celladdr checks for typeseq - P2
Now the ALphaNum type identifier also returns a list of types
which make up the AlphaNum token.

Inturn get_celladdr is updated to get a collapsed list of typeSeq
and then check if it matchs the pattern expected for CellAddresses.

TODO: Currently cell addresses with $ (to notify that that part of
cell address should be fixed) in them are not handled.