celladdr_valid using re - p1
SO that we dont forget there is also the regular expressions.

This is the 1st step in refining the re. Extract alpha and numeric
part of the cell address, by ignoring any $ inbetween.

TODO: It doesnt bother if the alpha part is too large.
TODO: It doesnt bother if there is any garbage following numeric part.

NOW $ in cell address, will be silently ignored and during cell address
updates during insertion and deletion, the $ will be silently dropped.