Update get_token default addon set
PReviously the get_token was considering only . as part of the
alphanum token by default. Now even $ is considered as part of
the alphanum token by default.

THis allows us to use $ as part of cell addresses, to indicate
to the program that - that part of the cell address shouldnt be
automatically adjusted when inserting or deleting rows/cols.

TODO: Update other parsing helper routines to taken ANAddOn
from its users, and inturn pass it to its get_token call. And inturn
maybe avoid this ANAddOnDefault set.