shell command - cstart cend no use, to revert
Commiting here to just capture this flow. But otherwise cstart
and cend doesnt seem to solve the problem which I noted in last
commit.

Also reading the man page of reset_shell_mode, it would appear
that def_prog|shell_mode is automatically done by curses on initscr.

So need to revert this commit and also remove the def_prog|shell_mode
calls in cui.cstart.