shell cmd: no need for def_shell|prg_mode
As noted in the commit that was reverted, and as noted in the man
page curses initscr already takes care of calling def_shell|prg_mode

So me calling it again in cui.cstart was what created the avoidable
mess of shell mode.

So now I have removed those from cui.cstart.

Also have added a redrawing of the program screen when returning
back into the program.

README and help.csv updated.