Remove the hardcoded quote char in logic
When converting the logic to pick the current quote char from a
global variable, instead of hardcoding in the code, I had missed
out doing this in the save_file logic partially.

Inturn as I had changed the default text quote char from ` to ',
this was leading to the corner case of cells with fieldsep as part
of their content being double quoted over two load-save cycles.
And then this leading to messup when the file is loaded, because
the quote cancels its self out and the quote protection for content
with fieldsep evaporates.

This has been fixed by removing the hardcoding of the old quote
char in save_file logic.