Trap calc loops - P1
When trying to fetch cell data to print on screen, if its a =expression,
then clear a count table.

Then as part of calculating this data, each time a cell value is fetched,
a counter is incremented wrt that particular cell.

If the count of any cell goes beyond twice the number of non empty cells
in the spreadsheet. It would mean with a high degree of probability, that
we are in a calc recursive dead loop.

In such a situation raise a exception.