Trap Calc Looping - T2
When a calc looping is trapped, reset the data in the corresponding
cell to contain a Error Tag, so that it will no longer be calced.

TODO: The above sets the error tag to the cell whose calc crossed
the looping threshold. However the actual looping expression could
belong to some other cell, which inturn is also refering to this
cell in the same looping expression. In which case this in itself
wont be sufficient. WIll look at fixing it next.