More aggressive trap calc loop
Instead of trying to wait for same cell being recalced

	NumOfNonEMptyCells * 10 times

Now it only waits for

	NumOfNonEmptyCells * 4 times

because otherwise in many cases the recursive depth is exceeded
much before my trap traps things.

TODO: Rather I should check if the calcd threshold to trigger trap
is too large and if so limit it. Maybe dont allow any cell to be
recalced more than say 50 times, as part of trying to find the
value of a =expression.

And or additionally dont allow total number of calcs to exceed 200
calcs in a single =expression calc.

And or Maybe more intelligent recursion (indirect) identification
logic, like

Have a recursion depth argument passed around as part of nvalue and
related functions and inturn check its depth to identify looping
and or too much depth related issues.

or ...