prefix m to readonly/readwrite mode commands
As there is a bunch of explicit commands which operate on a range
of cells which start with r.

So to avoid unnecessary confusion, if any. Have changed the readonly
and readwrite commands as mreadonly/mro and mreadwrite/mrw i.e
mode [readonly/readwrite].

Even the commandline argument to put program in readonly mode, has
been changed to --mreadonly