New take at CalcLooping Trapping using callDepth
This is the initial take at calc looping if any by using callDepth.

When ever solving of a =expression leads into a cell address to be
fetched, a global callDepth is incremented. And inturn when one
returns back by fetching the value, the callDepth is decremented.

Inturn use this callDepth to decide have we gone too much deep or not.