Set ErrTag to all cells involved in a CalcLoop
Currently all cells involved in the chain of calculations, which
led to the calc loop are err tag set.

TODO: In future, it can be done in a bit more intelligent way.
The lowest hanging fruit is to not set err tag for cells whose
=expression just contains a numeric value or expression with no
reference to any cell addresses.