trapCalcLoopSets cdataUpdate, cdraw_data use cdata
As trapCalcLoop error tags cells involved in a trapped calc looping,
so the cached data needs to be updated, so the flag is setup for same.

cdraw_data has been updated to use the cached data rather than the
raw data.

This speeds up the display of spreadsheets with huge calculations, as
there is no longer a need to recalculate the data for each and every
cell when the spreadsheet is being displayed to user.

Rather the calculation now occurs only when content of some cell is
changed due to some user action.