Trap value eval in cdata_update, so update all cells
cdata_update traps any exception raised when evaluating the =expression
of a given cell. This ensures that all cells get evaluated now, rather
than cells beyond the exception cell getting skipped if exception occurs.

So if there are trapCalcLoops in the spreadsheet, all of them get flags
in one go.