Update cell calculation cache from nvalue
Previously only cdata_update was allowed to update the cell calc
cache, while nvalue would only read from it.

Now even nvalue updates the cell calc cache, if a requested cell
is not yet available in cache and nvalue had to calculate it.

This helps speedup/optimise calculation in spreadsheets, where
the cells have dependencies on cells which are to their right.

This is because, PREVIOUSLY

cdata_update on its own would have only updated cache of cell
calculations by going through cells from

	left to right AND
	top to bottom

This would allow cells which depend on cells to its top and or left
reuse the cached calculated values of those cells. While dependencies
which are to the right will need to be calculated again and again.

NOW with nvalue also updating cdata, as and when a cell is calculated,
irrespective of from where the request came, each cell is inturn
calculated only once. Thus cell calculations are optimised at a base
level.

NOTE: THe Cell Calc caching also helps avoid/bypass deep cell-chaining
in many cases, especially if cells are dependent on cells to their left
on their own row and or any (to left or right doesnt matter) cell on
the rows to their top.

TODO: Need to use the data viewport to limit the amount of cdata_update
at any given time. This concept was what was done originally, when each
cell was calculated each time it was required. Need to apply the same
to the cached cell calc flow also, so that when loading a spreadsheet
and or modifying a cell, only required cells are (re)calculated rather
than all cells in the spreadsheet.