Recursion Error eaten up by re, need diff strategy
Trying to trap recursion error in _nvalue also doesnt help.

At one level it turned out as expected, as the program was dumping
recursion error somewhere in python re and re seems to be trapping
the same and not re-raising it.

So the recursion error trapped recovery of the cdata_update wont
work out. Will be trying a different strategy next.