cdraw_data do partial cdata_update [R6/N]
Do cell calculations only for cells within the data window/viewport.
This is to help handle spreadsheets with hell lot of calculations
involving very very long chaining of cells. This helps reduce the
time taken to

* load such spreadsheet

* edit such spreadsheets, for cells not in the chain and away from
  view involving the heavy hitting cells.

NOTE: However when one has to show the cells involving the heavy
chaining (in the direction opposite to the flow of cell calculations)
and or edit them or cells around them, the program does take time
to do the needed calculations. Once calculated, if there is no
changes being done to cells, but only navigation, then the performance
will be smooth, as the cdata cell calc cache will come into play.

NOTE: SystemCRASH .gitCorruptionRecovery (workDir, .gitManualScavnging)