Put a status msg wrt cdata_update running
Pass current curses screen/window handle to cdraw_data.

Use it to draw a status message at the bottom right corner wrt
possible processing wrt cdata_update.

THis ensures that if there is a long calculation that needs to
occur, then the user will see this status message about processing
and knows that something is occuring when the calculations are
occuring in the background.

This is mainly useful for cases with lot of deep cell-chaining
calculations, with the cell-dependence opposite to the programs
left-to-right and top-to-bottom cell calc flow.