Num2ColAddr Account for OffBy1 At Each new Digit level
The Alphabetic numbering in spreadsheets have a off by 1 characteristic,
for each digit|char level in the address. Account for same.

In testing proper handling of same both in Num2Alpha and ALpha2Num
conversion path, by creating a simple spreadsheet which adds two numbers
in the cols AZZ and BAA from column A, noticed that libreoffice doesnt
seem to support column more than AMZ or so.