cfieldsep No need to replace fieldsep in text cells.
As existing fieldsep in text cell contents, has nothing to do with
the fieldsep of the csv file logically, and is potentially there for
some other reason. SO there is no need (and it is even wrong) to
replace old fieldsep with new fieldsep within text cell content,
when cfieldsep is triggered by user.

ANd eitherway while saving, current fieldsep, is found to be part
of text cell contents, then it will get automatically quoted.