Make main dlg and status with row=0 work always.
main's dlg and status was previously using cellstr from main logic,

This inturn meant that if one passes a msg list with multiple strings,
then only 1st message was guarenteed to be visible (provided the row
passed was 0), while subsequent messages would show up only if the rows
currently visible included rows to which the subsequent messages would
map to, based on the starting row passed to the function.

This was because the arguments to dlg/status and inturn cellstr was
based on data cell matrix row and col. And 0th row will always be on
screen while other cell matrix rows may or maynot be visible on the
screen.

TO avoid this corner case, now main's dlg and status get the position
of the row passed to it, and from then on increment the screen y position,
for subsequent messages in the list.

So if row is 0, then all messages will be visible. And in other
cases, based on the where the data row corresonding to the row number
passed to dlg/status is, the messages could be visible or not.

NOTE: Because main's dlg/status is based on data cell matrix row,col
positioning, so if the corresponding data row,col is currently
not visible then even this dlg/status wont be visible.