Simplify cdata_update as nvalue_key does sane things
nvalue_key takes care of interpreting the contents of a cell as
required, where numeric value is prefered.

Also it takes care of adding the result to cache, if not already
in cache.

so cdata_update simplified to depend on nvalue_key to do the right
thing. THus avoiding duplicating of logic to interpret cell contents.

NOTE: Now interpretation of cell contents is consolidated and
logic blackholed into nvalue_key. All other logics depend on it,
be it value_key, nvalue_saddr_or_str or cdata_update or ...