GBFlyPython: Bit more
As a functions default argument value is locked in when the module
containing the function was loaded into python. So trying to change
GBText2Zero based on flypython cmdline argument and inturn have it
have effect wrt nvalue_key doesnt work out.

So now nvalue_key sets its bText2Zero func argument to None by default
and inturn if None, picks up the current value of GBText2Zero, when
the func is called.

This allows one to put a simple python statement|expression in a cell
and inturn run the same by refering to that cell from another cell.

Also updated the test file to show a bit more of same.