Make rgennums insert direct numeric rather than =numeric
As the program now allows numbers to be entered into cells directly,
so use the same in rgennums.

This inturn ensures that the number of expressions that need to be
checked when inserting or deleting rows/cols, reduces by a good
amount and thus speed up the operations.

NOTE: in general insert and delete need to check each =expression to
see if they have any cell addresses in them that need adjusting or
not.

OPTI:TODO: parse =expressions and mark those which have cell addresses,
so that only they get checked during insert/delete as well as during
recalculation of cells.

Have two filled csv test files. One using =expressions and another
using direct numbers.

General README updates also.