PathCompletion: Finetune basename matching
NOTE: Had forgotten to look at the ErrLog which also captures exceptions.
SO wasted a lot of time, assuming the basic flow was failing, when in
reality, I had forgotten to set listSub in the main path, so the subpaths
were never triggering, instead were raising a exception.

With the listSub initialised, the basic flow is working at one level.

However if there is a sublist of items with the same prefix as one of the
items in that sublist, then it doesnt exit that list and go to next items
in the main list, in all cases, need to fix this.