PathCompletion: Try exit subList looping
TO help with same, use sorted lists. Because otherwise the filtered
sublist could turn out to be longer than a distributed set of sublists
with same prefix in the main list. IN which case repositioning when
looping within sublist is identified will become difficult.

Also from a end user perspective, it is more friendly to tab through
sorted list of possible paths.