PathCompletion: when curString is only what user entered
THe previous path completion logic was meant to handle the case,
where the current string passed to the program could either include

	what user had entered and or

	what the logic had returned to user previously,
	even if user hasnt indicated any preference for this path.

So it was implemented such that it allows any sub set of path basenames
with same prefix (which is also a valid basename) dont get stuck in a
loop / backhole of its own for ever, so it has a mechanism to exit such
subgroups of basenames after n loops.

But now given that sCur will only contain what user has entered or
selected from previous suggestions, on their own, so no need to worry
if the user selection (i.e sCur) locks the path completion logic into
a subset of all possible basenames, because that is what user has
requested.

So that simplifies the logic to a good extent. Also this leads to a
more predictable and controllable tab completion logic from user
perspective.