Another Stab at TabCOmplete
Moved the logic which updates the lists involved be it the main
list or the sublist into a seperate function.

This is because the sublist needed updating at two different locations

	one where main list is valid and sublist needs creating

	another where main list is invalid and thus both main and
	sublist needs creating.

Also the creation logic itself returns what to return back to user,
if lists have been created/updated.

If there is no matching things then return the input back.