cellContentUpdate: CopyOfOrigFwdLink, =Expressions or not
Maintain a copy of the original fwdLink so that after the new =exp
has been parsed, one can check if any prev cell addresses have been
dropped or not. This needs to be used to drop the revLinks to the
current cell from those dropped cells.

Crosscheck if it is a =expression, if not dont parse for cell addrs.