Sync up after a file is loaded
When a file is loaded, create the fwd and rev links as required
for the full spreadsheet which has been loaded, So that the program
is aware of all the interdependencies between the cells and inturn
can update related cell calcs when cell contents are edited.

TODO: Need to sync the links when rows/cols are inserted and or
deleted.