Handle Insert op and calc cache data
If the insert operation doesnt generate any err tag for a given
cell after the insert operation, then its cdata is copied into
the equivalent location (same or different), as is, because
insert doesnt change the value of cells normally.

Inturn the cdataUpdate is no longer set to True, because the cdata
is adjusted as required, including where there may be a issue, the
cdata is not set for corresponding cell in the new cdata dictionary.